requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
lpips>=0.1.4
pytorch-msssim>=0.2.1
opencv-python>=4.5.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
tqdm>=4.65.0 